Apply the default error handler to every API, and marshal as Error
