+++ archive/archive_api/src/archive_api.py
@@ -8,6 +8,7 @@ from flask_restplus import Api
 
 from apis import bags_api, ingests_api
 import config
+import models
 from responses import ContextResponse
 from progress_manager import ProgressManager
 
@@ -39,6 +40,29 @@ app.config["PROGRESS_MANAGER"] = ProgressManager(
 
 api.add_namespace(bags_api)
 api.add_namespace(ingests_api)
+
+
+# TODO: There's no testing of the error handling; we should fix that!
+@app.errorhandler(Exception)
+@api.errorhandler(Exception)
+@bags_api.errorhandler(Exception)
+@ingests_api.errorhandler(Exception)
+@api.marshal_with(models.Error, skip_none=True)
+def default_error_handler(error):
+    error_response = {
+        "httpStatus": getattr(error, "code", 500),
+        "label": getattr(error, "name", "Internal Server Error"),
+    }
+    logger.warn(error)
+    if error_response["httpStatus"] != 500:
+        if hasattr(error, "data"):
+            error_response["description"] = ", ".join(
+                error.data.get("errors", {}).values()
+            )
+        else:
+            error_response["description"] = getattr(error, "description", str(error))
+    return error_response, error_response["httpStatus"]
+
 
 # We can't move this import to the top because the views need the ``api``
 # instance defined in this file.
